tweak psh clone script to ignore database dump+import of empty entries

+++ scripts/platform_sh_clone.sh
@@ -80,19 +80,19 @@ fi
 echo "> Fetch data."
 PLATFORM_RELATIONSHIPS=`ssh $SSH_URL 'echo $PLATFORM_RELATIONSHIPS | base64 -d'`
 PLATFORM_VARIABLES=`ssh $SSH_URL 'echo $PLATFORM_VARIABLES | base64 -d'`
-CONFIG_JSON=`pcc project:configjson`
+CONFIG_JSON=$(${PCC_PATH} project:configjson)
 
 # - set variables
 echo "> Set variables."
 VAR_LIST=$(get_variables "$PLATFORM_VARIABLES")
 while IFS= read -r var; do
     IFS=';' read -ra var_split <<< "$var"
-    pcc var:set "${var_split[0]}" "${var_split[1]}"
+    $PCC_PATH var:set "${var_split[0]}" "${var_split[1]}"
 done <<< "$VAR_LIST"
 
 # - start project
 echo "> Start project."
-pcc project:start
+$PCC_PATH project:start
 
 # - get databse dumps (mysql)
 echo "> Fetch database dumps."
@@ -100,13 +100,15 @@ DATABASE_LIST=$(get_databases "$PLATFORM_RELATIONSHIPS")
 while IFS= read -r db; do
     IFS=';' read -ra db_vals <<< "$db"
     IFS=
-    echo "> Dump ${db_vals[4]}."
-    ssh $SSH_URL "mysqldump --host=${db_vals[0]} --port=${db_vals[1]} --user=${db_vals[2]} --password=${db_vals[3]} ${db_vals[4]}" | gzip > /tmp/dump.sql.gz
-    gunzip /tmp/dump.sql.gz
-    echo "> Import ${db_vals[4]}."
-    echo "drop schema if exists ${db_vals[4]}; create schema ${db_vals[4]}" | pcc mysql:sql
-    $PCC_PATH mysql:sql -d "${db_vals[4]}" < /tmp/dump.sql
-    rm /tmp/dump.sql    
+    if [ ! -z "${db_vals[4]}" ]; then
+        echo "> Dump ${db_vals[4]}."
+        ssh $SSH_URL "mysqldump --host=${db_vals[0]} --port=${db_vals[1]} --user=${db_vals[2]} --password=${db_vals[3]} ${db_vals[4]}" | gzip > /tmp/dump.sql.gz
+        gunzip /tmp/dump.sql.gz
+        echo "> Import ${db_vals[4]}."
+        echo "drop schema if exists ${db_vals[4]}; create schema ${db_vals[4]}" | $PCC_PATH mysql:sql
+        $PCC_PATH mysql:sql -d "${db_vals[4]}" < /tmp/dump.sql
+        rm /tmp/dump.sql
+    fi
 done <<< "$DATABASE_LIST"
 
 # - upload ssh key to app
